pyproject.toml
[project]
name = "ldap_idp"
version = "0.1.0"
description = "Simple ldap-manager"
authors = [
    {name = "mrjk"}
]
license = {text = "GPLv3"}
readme = "README.md"
requires-python = ">3.12,<4"
dependencies = [
    "textual (>=4.0.0,<5.0.0)",
    "python-ldap (>=3.4.4,<4.0.0)",
    "click (>=8.0.0,<9.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "textual-serve (>=1.1.2,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[project.scripts]
ldap-idp = "ldap_idp.main:main"

[tool]
[tool.poetry]
[tool.poetry.group.dev]
[tool.poetry.group.dev.dependencies]
textual-dev = "^1.7.0"
black = "^25.1.0"
pytest = "^8.4.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
